Fixed "conversion from 'double' to 'int', possible loss of data"
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@7939 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/script.c
@@ -2379,7 +2379,7 @@ void op_2num(struct script_state *st,int op,int i1,int i2)
 		if(ret_double > INT_MAX || ret_double < INT_MIN) {
 			printf("script::op_2num overflow detected op:%d\n",op);
 			report_src(st);
-			ret = cap_value(ret_double,INT_MAX,INT_MIN);
+			ret = (int)cap_value(ret_double,INT_MAX,INT_MIN);
 		}
 	}
 	push_val(st->stack,C_INT,ret);
